feat: Implement restaurant menu feature and refactor components
- Added RestaurantMenuCard component to display menu items for selected restaurant.
- Integrated useRestaurantMenu hook to fetch and manage restaurant menu data.
- Updated routing in App.js to include restaurant-specific paths.
- Refactored Body component to fetch restaurant data from API instead of using hardcoded list.
- Created mock data file for restaurant list to facilitate testing.
- Enhanced RestaurantCard component to link to individual restaurant menus.
- Updated constants.js to include new API endpoints for fetching restaurant menus.
