+++ src/App.js
@@ -5,6 +5,7 @@ import Body from "./components/Body";
 import AboutUs from "./components/AboutUs";
 import Contact from "./components/Contact";
 import Error from "./components/Error";
+import RestaurantMenuCard from "./components/RestaurantMenuCard";
 import {createBrowserRouter, RouterProvider,Outlet} from "react-router-dom";
 
 const Instamart = React.lazy(() => import("./components/Instamart"));
@@ -38,6 +39,10 @@ const appRouter = createBrowserRouter([
             <Instamart />
           </React.Suspense>
         ),
+      },
+      {
+        path: "/restaurants/:resId",
+        element: <RestaurantMenuCard />,
       },
       {
         path: "/contact",
