+++ src/utils/mockData/resList.js
@@ -1,0 +1,509 @@
+const restList = [
+  {
+    info: {
+      id: "38925",
+      name: "Domino's Pizza",
+      cloudinaryImageId:
+        "RX_THUMBNAIL/IMAGES/VENDOR/2025/6/5/9b2b8803-ff5b-4578-9e91-6c0fd2efd5af_38925.JPG",
+      locality: "Netaji Subhash Marg",
+      areaName: "Daryaganj",
+      costForTwo: "₹400 for two",
+      cuisines: ["Pizzas", "Italian", "Pastas", "Desserts"],
+      avgRating: 4.3,
+      parentId: "2456",
+      avgRatingString: "3.9",
+      totalRatingsString: "12K+",
+      promoted: true,
+      adTrackingId:
+        "cid=29942659~p=1~adgrpid=29942659#ag1~mp=SWIGGY_IN~bl=FOOD~aet=RESTAURANT~aeid=38925~plpr=COLLECTION~eid=46be93e3-f3d8-4cee-9430-bce34972c5d6~srvts=1749291906284~collid=115117",
+      sla: {
+        deliveryTime: 25,
+        lastMileTravel: 4.6,
+        serviceability: "SERVICEABLE",
+        slaString: "20-25 mins",
+        lastMileTravelString: "4.6 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2025-06-07 23:59:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "ITEMS",
+        subHeader: "AT ₹89",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+      externalRatings: {
+        aggregatedRating: {
+          rating: "--",
+        },
+      },
+      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
+    },
+    analytics: {
+      context: "seo-data-a8213097-a7fe-4a6d-ad48-beeeba0449d3",
+    },
+    cta: {
+      link: "https://www.swiggy.com/city/delhi/dominos-pizza-netaji-subhash-marg-daryaganj-rest38925",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "908488",
+      name: "Biryani Blues",
+      cloudinaryImageId:
+        "RX_THUMBNAIL/IMAGES/VENDOR/2024/6/12/98a7e169-4ffd-41c1-ba08-ccb564af9abb_908488.jpg",
+      locality: "Chandni Chowk",
+      areaName: "Old Delhi",
+      costForTwo: "₹400 for two",
+      cuisines: ["Biryani", "Hyderabadi"],
+      avgRating: 4.2,
+      parentId: "13813",
+      avgRatingString: "3.9",
+      totalRatingsString: "374",
+      sla: {
+        deliveryTime: 40,
+        lastMileTravel: 1.4,
+        serviceability: "SERVICEABLE",
+        slaString: "35-40 mins",
+        lastMileTravelString: "1.4 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2025-06-07 23:59:00",
+        opened: true,
+      },
+      badges: {
+        imageBadges: [
+          {
+            imageId: "Rxawards/_CATEGORY-Biryani.png",
+            description: "Delivery!",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "Delivery!",
+                  imageId: "Rxawards/_CATEGORY-Biryani.png",
+                },
+              },
+            ],
+          },
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹100 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+      externalRatings: {
+        aggregatedRating: {
+          rating: "--",
+        },
+      },
+      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
+    },
+    analytics: {
+      context: "seo-data-a8213097-a7fe-4a6d-ad48-beeeba0449d3",
+    },
+    cta: {
+      link: "https://www.swiggy.com/city/delhi/biryani-blues-chandni-chowk-old-delhi-rest908488",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "350774",
+      name: "Subway",
+      cloudinaryImageId:
+        "RX_THUMBNAIL/IMAGES/VENDOR/2025/1/21/5a7a502c-3f31-4b93-b026-b858b402bb72_350774.JPG",
+      locality: "Satanam Road",
+      areaName: "Krishna Nagar",
+      costForTwo: "₹350 for two",
+      cuisines: ["sandwich", "Salads", "wrap", "Healthy Food"],
+      avgRating: 4.1,
+      parentId: "2",
+      avgRatingString: "4.1",
+      totalRatingsString: "4.1K+",
+      promoted: true,
+      adTrackingId:
+        "cid=29737895~p=2~adgrpid=29737895#ag1~mp=SWIGGY_IN~bl=FOOD~aet=RESTAURANT~aeid=350774~plpr=COLLECTION~eid=9b15e8e8-6739-43ce-a636-b3e893fa65a3~srvts=1749291906284~collid=115117",
+      sla: {
+        deliveryTime: 31,
+        lastMileTravel: 5,
+        serviceability: "SERVICEABLE",
+        slaString: "30-35 mins",
+        lastMileTravelString: "5.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2025-06-08 02:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "20% OFF",
+        subHeader: "UPTO ₹60",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+      externalRatings: {
+        aggregatedRating: {
+          rating: "--",
+        },
+      },
+      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
+    },
+    analytics: {
+      context: "seo-data-a8213097-a7fe-4a6d-ad48-beeeba0449d3",
+    },
+    cta: {
+      link: "https://www.swiggy.com/city/delhi/subway-satanam-road-krishna-nagar-rest350774",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "958130",
+      name: "Burger Singh(Big Punjabi Burgers)",
+      cloudinaryImageId:
+        "RX_THUMBNAIL/IMAGES/VENDOR/2024/11/25/02f6bd53-7c7b-48e7-a102-f1f577ab87ea_958130.jpg",
+      locality: "KOTWALI",
+      areaName: "Pratap Nagar",
+      costForTwo: "₹300 for two",
+      cuisines: [
+        "Burgers",
+        "Snacks",
+        "Desserts",
+        "Beverages",
+        "Fast Food",
+        "American",
+        "Indian",
+        "Biryani",
+      ],
+      avgRating: 4.2,
+      parentId: "375065",
+      avgRatingString: "3.8",
+      totalRatingsString: "669",
+      sla: {
+        deliveryTime: 32,
+        lastMileTravel: 4.9,
+        serviceability: "SERVICEABLE",
+        slaString: "30-35 mins",
+        lastMileTravelString: "4.9 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2025-06-08 06:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "ITEMS",
+        subHeader: "AT ₹47",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+      externalRatings: {
+        aggregatedRating: {
+          rating: "--",
+        },
+      },
+      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
+    },
+    analytics: {
+      context: "seo-data-a8213097-a7fe-4a6d-ad48-beeeba0449d3",
+    },
+    cta: {
+      link: "https://www.swiggy.com/city/delhi/burger-singh-big-punjabi-burgers-kotwali-pratap-nagar-rest958130",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "914078",
+      name: "California Burrito",
+      cloudinaryImageId:
+        "RX_THUMBNAIL/IMAGES/VENDOR/2024/9/3/35fc7ea0-580c-4827-855a-1878e03f3796_914078.jpg",
+      locality: "Kamla Nagar",
+      areaName: "GTB Nagar",
+      costForTwo: "₹250 for two",
+      cuisines: ["American", "Salads"],
+      avgRating: 4.5,
+      parentId: "432183",
+      avgRatingString: "4.5",
+      totalRatingsString: "1.2K+",
+      promoted: true,
+      adTrackingId:
+        "cid=4fde2a53-15eb-4bce-a26f-1b3f6bab6423~p=4~adgrpid=4fde2a53-15eb-4bce-a26f-1b3f6bab6423#ag1~mp=SWIGGY_IN~bl=FOOD~aet=RESTAURANT~aeid=914078~plpr=COLLECTION~eid=1e328e29-ad0d-4084-af3e-1a16edcac6b4~srvts=1749291906284~collid=115117",
+      sla: {
+        deliveryTime: 34,
+        lastMileTravel: 6.1,
+        serviceability: "SERVICEABLE",
+        slaString: "30-35 mins",
+        lastMileTravelString: "6.1 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2025-06-07 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      aggregatedDiscountInfoV2: {},
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+      externalRatings: {
+        aggregatedRating: {
+          rating: "--",
+        },
+      },
+      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
+    },
+    analytics: {
+      context: "seo-data-a8213097-a7fe-4a6d-ad48-beeeba0449d3",
+    },
+    cta: {
+      link: "https://www.swiggy.com/city/delhi/california-burrito-kamla-nagar-gtb-nagar-rest914078",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "914832",
+      name: "Thalairaj Biryani",
+      cloudinaryImageId:
+        "RX_THUMBNAIL/IMAGES/VENDOR/2024/7/25/8922aa1d-0b4b-4800-ab30-3faa8c28113f_914832.jpg",
+      locality: "Kamla Nagar",
+      areaName: "GTB Nagar",
+      costForTwo: "₹250 for two",
+      cuisines: ["Andhra", "Biryani", "South Indian", "Kebabs"],
+      avgRating: 4.5,
+      parentId: "433875",
+      avgRatingString: "4.5",
+      totalRatingsString: "768",
+      sla: {
+        deliveryTime: 35,
+        lastMileTravel: 6.1,
+        serviceability: "SERVICEABLE",
+        slaString: "35-40 mins",
+        lastMileTravelString: "6.1 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2025-06-07 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹200 OFF",
+        subHeader: "ABOVE ₹999",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+      externalRatings: {
+        aggregatedRating: {
+          rating: "--",
+        },
+      },
+      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
+    },
+    analytics: {
+      context: "seo-data-a8213097-a7fe-4a6d-ad48-beeeba0449d3",
+    },
+    cta: {
+      link: "https://www.swiggy.com/city/delhi/thalairaj-biryani-kamla-nagar-gtb-nagar-rest914832",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "735722",
+      name: "ENSO - Sourdough Pizza by Nomad",
+      cloudinaryImageId:
+        "RX_THUMBNAIL/IMAGES/VENDOR/2025/1/9/8a800bf4-1432-418e-ad55-dd1d2d81aefe_735722.jpg",
+      locality: "Jain Mandir Marg",
+      areaName: "Connaught Place",
+      costForTwo: "₹850 for two",
+      cuisines: ["Pizzas", "Fast Food", "Desserts", "Salads", "Beverages"],
+      avgRating: 4.1,
+      parentId: "77344",
+      avgRatingString: "4.1",
+      totalRatingsString: "236",
+      sla: {
+        deliveryTime: 43,
+        lastMileTravel: 7.4,
+        serviceability: "SERVICEABLE",
+        slaString: "40-45 mins",
+        lastMileTravelString: "7.4 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2025-06-08 05:29:00",
+        opened: true,
+      },
+      badges: {
+        imageBadges: [
+          {
+            imageId: "newg.png",
+            description: "Gourmet",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "Gourmet",
+                  imageId: "newg.png",
+                },
+              },
+            ],
+          },
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹175 OFF",
+        subHeader: "ABOVE ₹1249",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+      externalRatings: {
+        aggregatedRating: {
+          rating: "--",
+        },
+      },
+      ratingsDisplayPreference: "RATINGS_DISPLAY_PREFERENCE_SHOW_SWIGGY",
+    },
+    analytics: {
+      context: "seo-data-a8213097-a7fe-4a6d-ad48-beeeba0449d3",
+    },
+    cta: {
+      link: "https://www.swiggy.com/city/delhi/enso-sourdough-pizza-by-nomad-jain-mandir-marg-connaught-place-rest735722",
+      type: "WEBLINK",
+    },
+  },
+];
+
+export default restList;
